Add MainWithInMemory.java with in-memory DAO setup
